// src/api/api.generated.ts
export enum DTOCryptoCurrency {
  DTO_TON = "TON",
  DTO_USDT = "USDT",
}

/** @default "mainnet" */
export enum DTOChain {
  DTOMainnet = "mainnet",
  DTOTestnet = "testnet",
}

/** @example "pending" */
export enum DTOInvoiceStatus {
  DTOPending = "pending",
  DTOPaid = "paid",
  DTOCancelled = "cancelled",
  DTOExpired = "expired",
}

export enum DTOServiceName {
  DTOMessages = "messages",
  DTOStats = "stats",
  DTOTonapi = "tonapi",
  DTOTestnet = "testnet",
  DTOCnft = "cnft",
  DTOLiteproxy = "liteproxy",
  DTOStreaming = "streaming",
}

export enum DTOStatsQueryType {
  DTOGraph = "graph",
  DTOBaseQuery = "base_query",
  DTOChatGptQuery = "chat_gpt_query",
}

export enum DTOStatsQueryStatus {
  DTOExecuting = "executing",
  DTOSuccess = "success",
  DTOError = "error",
}

export enum DTOTokenCapability {
  DTOWebhooks = "webhooks",
  DTOCocoon = "cocoon",
}

export type DTOBillingTxInfo =
  | ({
      reason: "tonapi_monthly_payment";
    } & any)
  | ({
      reason: "tonapi_instant_payment";
    } & any)
  | ({
      reason: "tonapi_change_tier";
    } & any)
  | ({
      reason: "liteproxy_monthly_payment";
    } & any)
  | ({
      reason: "liteproxy_change_tier";
    } & any)
  | ({
      reason: "testnet_tons_purchase";
    } & any)
  | ({
      reason: "cnft_indexing_payment";
    } & any)
  | ({
      reason: "message_package_purchase";
    } & any)
  | ({
      reason: "chatgpt_request_payment";
    } & any)
  | ({
      reason: "promo_code_activation";
    } & any)
  | ({
      reason: "streaming_api_payment";
    } & any)
  | ({
      reason: "cocoon_api_payment";
    } & any)
  | ({
      reason: "replenishment_of_deposit";
    } & any)
  | ({
      reason: "analytics_request_payment";
    } & any)
  | ({
      reason: "other";
    } & any);
// ...
export interface DTOAppTier {
  /**
   * @format uint32
   * @example 1
   */
  id: number;
  /** @example "Test tier" */
  name: string;
  /** @example 5 */
  rpc: number;
  /** @example 100 */
  usd_price: number;
  /** @example 50 */
  long_polling_sub: number;
  /** @example 10 */
  entity_per_conn: number;
  capabilities: string[];
  /**
   * @format int64
   * @example 1690889913000
   */
  next_payment?: number;
  /**
   * @format int64
   * @example 1690889913000
   */
  date_create: number;
  /** @example false */
  instant_payment: boolean;
}

export interface DTOBillingTransaction {
  /**
   * @format uuid
   * @example "03cfc582-b1c3-410a-a9a7-1f0afe116b3b"
   */
  id: string;
  type: DTOBillingTransactionTypeEnum;
  /**
   * deposited amount
   * @example "123000000000"
   */
  amount: string;
  currency: DTOCryptoCurrency;
  info: DTOBillingTxInfo;
  description: string;
  /**
   * @format int64
   * @example 1690889913000
   */
  created_at: number;
}
// ...
export interface DTOBalance {
  /**
   * deposited amount
   * @example "123000000000"
   */
  amount: string;
  /** @example "12300000000" */
  promo_amount: string;
}

export interface DTOCharge {
  /** @example "742af494-e2cd-441f-98e8-ac6075280eff" */
  id: string;
  service: DTOServiceName;
  /**
   * @format uint32
   * @example 1
   */
  tier_id?: number;
  /**
   * @format uint32
   * @example 1
   */
  messages_package_id?: number;
  /**
   * @format uint32
   * @example 1
   */
  testnet_price_multiplicator?: number;
  /**
   * @format float64
   * @example 7.8
   */
  stats_spent_time?: number;
  /**
   * @format uint32
   * @example 1000000
   */
  stats_price_per_second?: number;
  stats_type_query?: DTOStatsQueryType;
  /**
   * @format int64
   * @example 1000000
   */
  cnft_indexing_price?: number;
  /**
   * @format uint32
   * @example 1000000
   */
  cnft_count?: number;
  /** @example "0:97146a46acc2654y27947f14c4a4b14273e954f78bc017790b41208b0043200b" */
  cnft_collections?: string;
  /**
   * @format int64
   * @example 1000000000
   */
  amount: number;
  /** @example 2.25 */
  exchange_rate: number;
  /**
   * @format int64
   * @example 1690889913000
   */
  date_create: number;
}
// ...
export interface DTOProjectLiteproxyTierDetail {
  /**
   * @format uint32
   * @example 1
   */
  id: number;
  /** @example "Test tier" */
  name: string;
  /** @example 5 */
  rps: number;
  /** @example 100 */
  usd_price: number;
  /**
   * @format int64
   * @example 1690889913000
   */
  next_payment?: number;
  /**
   * @format int64
   * @example 1690889913000
   */
  date_create: number;
}

/** backend error code */
export enum DTOErrorCodeEnum {
  DTOValue1 = 1,
  DTOValue2 = 2,
  DTOValue3 = 3,
}

export enum DTOBillingTransactionTypeEnum {
  DTODeposit = "deposit",
  DTOCharge = "charge",
}